# pkg/apidocs/openapi.yaml
openapi: 3.1.0
info:
  title: Candidate Take-Home API
  version: 1.0.0
  description: >-
    REST contract for assessment exercises A1, A2, and A3. WebSocket events are
    specified in API_SPEC.md. Tenant identity always comes from authentication.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /projects:
    get:
      tags: [Projects]
      operationId: listProjects
      summary: List tenant Projects
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Limit' }
        - name: search
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Paginated tenant Projects
          headers:
            X-Request-Id: { $ref: '#/components/headers/RequestId' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectPage' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Projects]
      operationId: createProject
      summary: Create a Project with nested Phases and Tasks atomically
      description: >-
        Admin only. Keys are scoped to the tenant. Repeat the same Idempotency-Key and body to
        receive the original result; reusing the key with a different body returns 409 IDEMPOTENCY_CONFLICT.
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema: { type: string, format: uuid, maxLength: 255 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateProject' }
      responses:
        '201':
          description: Project and all nested resources created
          headers:
            X-Request-Id: { $ref: '#/components/headers/RequestId' }
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { $ref: '#/components/schemas/Project' }
        '400':
          description: Missing Idempotency-Key header or malformed JSON
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /projects/{id}:
    parameters:
      - { $ref: '#/components/parameters/Id' }
    get:
      tags: [Projects]
      operationId: getProject
      summary: Get a tenant Project and its complete hierarchy
      responses:
        '200':
          description: Project
          headers:
            X-Request-Id: { $ref: '#/components/headers/RequestId' }
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { $ref: '#/components/schemas/Project' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Projects]
      operationId: deleteProject
      summary: Delete a tenant Project
      description: Admin only; used for assessment cleanup.
      responses:
        '200': { $ref: '#/components/responses/EmptySuccess' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  headers:
    RequestId:
      description: Request correlation ID.
      schema: { type: string }
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string }
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
  responses:
    EmptySuccess:
      description: Successful operation with no result body beyond the envelope.
      headers:
        X-Request-Id: { $ref: '#/components/headers/RequestId' }
      content:
        application/json:
          schema:
            type: object
            required: [data]
            properties:
              data: { type: 'null' }
    Unauthorized:
      description: Missing, invalid, expired, revoked, or reused credentials.
      headers:
        X-Request-Id: { $ref: '#/components/headers/RequestId' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Caller lacks the required role or tenant ownership.
      headers:
        X-Request-Id: { $ref: '#/components/headers/RequestId' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Resource does not exist in the caller's authorized scope.
      headers:
        X-Request-Id: { $ref: '#/components/headers/RequestId' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: Duplicate, stale version, or idempotency key payload conflict.
      headers:
        X-Request-Id: { $ref: '#/components/headers/RequestId' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    ValidationError:
      description: All discovered field validation errors.
      headers:
        X-Request-Id: { $ref: '#/components/headers/RequestId' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
...
    PageMeta:
      type: object
      required: [page, limit, total, totalPages]
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
        totalPages: { type: integer }
...
    Task:
      type: object
      required: [id, title, type, priority, assigneeId, estimateHours, dueDate, severity]
      properties:
        id: { type: string }
        title: { type: string }
        type: { type: string, enum: [feature, bug] }
        priority: { type: string, enum: [low, medium, high, critical] }
        assigneeId: { type: string }
        estimateHours: { type: number, exclusiveMinimum: 0, maximum: 999 }
        dueDate: { type: string, format: date }
        severity: { type: [string, 'null'], enum: [minor, major, critical, null] }
    CreateTask:
      type: object
      required: [title, type, priority, assigneeId, estimateHours, dueDate, severity]
      properties:
        title: { type: string, minLength: 1, maxLength: 200 }
        type: { type: string, enum: [feature, bug] }
        priority: { type: string, enum: [low, medium, high, critical] }
        assigneeId: { type: string }
        estimateHours: { type: number, exclusiveMinimum: 0, maximum: 999 }
        dueDate: { type: string, format: date }
        severity: { type: [string, 'null'], enum: [minor, major, critical, null] }
    Phase:
      type: object
      required: [id, name, order, startDate, endDate, tasks]
      properties:
        id: { type: string, readOnly: true }
        name: { type: string }
        order: { type: integer, minimum: 1 }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        tasks: { type: array, minItems: 1, maxItems: 100, items: { $ref: '#/components/schemas/Task' } }
    CreatePhase:
      type: object
      required: [name, order, startDate, endDate, tasks]
      properties:
        name: { type: string, minLength: 1, maxLength: 100 }
        order: { type: integer, minimum: 1, maximum: 1000, description: Unique within the Project. }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        tasks: { type: array, minItems: 1, maxItems: 100, items: { $ref: '#/components/schemas/CreateTask' } }
    Project:
      type: object
      required: [id, name, code, ownerId, startDate, endDate, phases, version, createdAt, updatedAt]
      properties:
        id: { type: string, readOnly: true }
        name: { type: string, minLength: 3, maxLength: 100 }
        code: { type: string }
        description: { type: [string, 'null'] }
        ownerId: { type: string }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        phases: { type: array, minItems: 1, maxItems: 20, items: { $ref: '#/components/schemas/Phase' } }
        version: { type: integer, minimum: 1 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CreateProject:
      type: object
      required: [name, code, ownerId, startDate, endDate, phases]
      properties:
        name: { type: string, minLength: 3, maxLength: 100 }
        code: { type: string, minLength: 1, maxLength: 50, description: Unique within the tenant. }
        description: { type: [string, 'null'], maxLength: 2000 }
        ownerId: { type: string }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        phases: { type: array, minItems: 1, maxItems: 20, items: { $ref: '#/components/schemas/CreatePhase' } }
    ProjectPage:
      type: object
      required: [data, meta]
      properties:
        data: { type: array, items: { $ref: '#/components/schemas/Project' } }
        meta: { $ref: '#/components/schemas/PageMeta' }
...
    ErrorResponse:
      type: object
      required: [error, requestId]
      properties:
        error:
          type: object
          required: [code, message, details]
          properties:
            code:
              type: string
              enum: [BAD_REQUEST, UNAUTHORIZED, TOKEN_EXPIRED, FORBIDDEN, NOT_FOUND, CONFLICT, VERSION_CONFLICT, IDEMPOTENCY_CONFLICT, VALIDATION_ERROR, RATE_LIMITED, INTERNAL_ERROR]
            message: { type: string }
            details:
              type: array
              items:
                type: object
                required: [field, message]
                properties:
                  field: { type: string }
                  message: { type: string }
        requestId: { type: string }